Some refactors to the commit package
The create commit func now returns a Commit struct instead of the commit
[]byte.

Adds a new method for Commit called ToBytes which returns the commit's
content as a byte slice.

Creates a new func in the commit package to write a commit struct to
ObjectDB.

Modifies the commands.Commit function to use these refactored and new
functions.